+++ tiger-common/src/main/java/de/gematik/test/tiger/common/data/config/tigerProxy/TigerProxyConfiguration.java
@@ -9,7 +9,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import de.gematik.rbellogger.modifier.RbelModificationDescription;
 import de.gematik.test.tiger.common.config.TigerConfigurationException;
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -35,8 +34,6 @@ public class TigerProxyConfiguration {
     @Builder.Default
     private TigerTlsConfiguration tls = TigerTlsConfiguration.builder().build();
     private List<String> keyFolders;
-    @Builder.Default
-    private boolean activateRbelEndpoint = false;
     @Builder.Default
     private boolean activateAsn1Parsing = false;
     @Builder.Default